# aspnetcore/src/IO.Swagger/Controllers/PaymentApi.cs
/*
 * Payment System API
 *
 * This is a simple API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: salimow.r2015@yandex.ru
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class PaymentApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="from"></param>
        /// <param name="to"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/Payment")]
        [ValidateModelState]
        [SwaggerOperation("ApiPaymentGet")]
        public virtual IActionResult ApiPaymentGet([FromQuery]DateTime? from, [FromQuery]DateTime? to)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="body"></param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/Payment")]
        [ValidateModelState]
        [SwaggerOperation("ApiPaymentPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(string), description: "Success")]
        public virtual IActionResult ApiPaymentPost([FromBody]PaymentInfoFromRequest body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(string));
            string exampleJson = null;
            exampleJson = "\"\"";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<string>(exampleJson)
                        : default(string);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}