# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Payment System API",
    "description" : "This is a simple API",
    "contact" : {
      "email" : "salimow.r2015@yandex.ru"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/Card" : {
      "post" : {
        "tags" : [ "Card" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CardInfoFromRequest"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CardInfoFromRequest"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CardInfoFromRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    },
    "/api/Payment" : {
      "get" : {
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "post" : {
        "tags" : [ "Payment" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentInfoFromRequest"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentInfoFromRequest"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentInfoFromRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/CreateAccount" : {
      "post" : {
        "tags" : [ "User" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewUser"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewUser"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/Login" : {
      "post" : {
        "tags" : [ "User" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginUser"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginUser"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginUser"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CardInfoFromRequest" : {
        "required" : [ "cardDate", "cardNumber", "sessionId", "verificationNumber" ],
        "type" : "object",
        "properties" : {
          "cardNumber" : {
            "maxLength" : 19,
            "type" : "string"
          },
          "verificationNumber" : {
            "maxLength" : 4,
            "minLength" : 3,
            "type" : "string"
          },
          "cardDate" : {
            "maxLength" : 5,
            "type" : "string"
          },
          "sessionId" : {
            "type" : "string"
          },
          "storeUrl" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "PaymentInfoFromRequest" : {
        "required" : [ "amount", "description" ],
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number",
            "format" : "double"
          },
          "description" : {
            "type" : "string"
          }
        },
        "additionalProperties" : false
      },
      "NewUser" : {
        "required" : [ "email", "name", "password" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "additionalProperties" : false
      },
      "LoginUser" : {
        "required" : [ "email", "password" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "additionalProperties" : false
      }
    }
  }
}